// superset-frontend/plugins/plugin-chart-unified-list-bar/src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import { QueryFormData, TimeseriesDataRecord } from '@superset-ui/core';

export interface UnifiedListBarChartStylesProps {
    height: number;
    width: number;
    headerFontSize?: string;
    boldText?: boolean;
}

export interface UnifiedListBarChartCustomizeProps {
    keyColumn: string;
    keySubColumn?: string;
    secondaryColumns: string[];
    metricColumn?: string;
    maxMetricColumn?: string;
    severityColumn?: string;
    colorColumn?: string;
    rowsPerItem: '1' | '2';
    alignMetric: 'left' | 'right';
    showBar: boolean;
    showMetricValue: boolean;
    keyFontSize: number;
    keyColor: string;
    keySubFontSize: number;
    secondaryFontSize: number;
    barColorPositive: string;
    barColorNegative: string;
    conditionalColorRules?: any[];
    iconRules?: any[];
}

export type UnifiedListBarChartQueryFormData = QueryFormData &
    UnifiedListBarChartStylesProps &
    UnifiedListBarChartCustomizeProps;

export interface UnifiedListBarChartProps extends UnifiedListBarChartStylesProps {
    data: TimeseriesDataRecord[];
    customize: UnifiedListBarChartCustomizeProps;
}